data_pipeline/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
tqdm==4.66.1
rottentomatoes-python==1.2.0
